requirements.txt
click==8.1.4
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.25.1
opencv-python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
